hmc5883: build on linux and qurt
