+++ boards/aerotenna/ocpoc/ubuntu.cmake
@@ -21,6 +21,7 @@ px4_add_board(
 		linux_pwm_out
 		linux_sbus
 		#magnetometer # all available magnetometer drivers
+		magnetometer/hmc5883
 		pwm_out_sim
 		#telemetry # all available telemetry drivers
