+++ boards/atlflight/eagle/qurt-default.cmake
@@ -48,6 +48,7 @@ px4_add_board(
 		barometer/bmp280
 		gps
 		imu/mpu9250
+		magnetometer/hmc5883
 		spektrum_rc
 		qshell/qurt
 		snapdragon_pwm_out
