+++ boards/emlid/navio2/cross.cmake
@@ -19,6 +19,7 @@ px4_add_board(
 		#imu # all available imu drivers
 		imu/mpu9250
 		#magnetometer # all available magnetometer drivers
+		magnetometer/hmc5883
 		pwm_out_sim
 		#telemetry # all available telemetry drivers
