+++ src/drivers/magnetometer/hmc5883/CMakeLists.txt
@@ -36,9 +36,11 @@ px4_add_module(
 	COMPILE_FLAGS
 		-Wno-cast-align # TODO: fix and enable
 	SRCS
+		HMC5883.cpp
+		HMC5883.hpp
 		hmc5883_i2c.cpp
 		hmc5883_spi.cpp
-		hmc5883.cpp
+		hmc5883_main.cpp
 	DEPENDS
 		px4_work_queue
 	)
