+++ src/drivers/magnetometer/hmc5883/hmc5883_i2c.cpp
@@ -37,27 +37,12 @@
  * I2C interface for HMC5883 / HMC 5983
  */
 
-/* XXX trim includes */
 #include <px4_platform_common/px4_config.h>
-
-#include <sys/types.h>
-#include <stdint.h>
-#include <stdbool.h>
-#include <string.h>
-#include <assert.h>
-#include <debug.h>
-#include <errno.h>
-#include <unistd.h>
-
-#include <arch/board/board.h>
-
 #include <drivers/device/i2c.h>
 #include <drivers/drv_mag.h>
 #include <drivers/drv_device.h>
 
 #include "hmc5883.h"
-
-#include "board_config.h"
 
 #define HMC5883L_ADDRESS		0x1E
