+++ src/drivers/magnetometer/hmc5883/hmc5883_spi.cpp
@@ -37,26 +37,12 @@
  * SPI interface for HMC5983
  */
 
-/* XXX trim includes */
 #include <px4_platform_common/px4_config.h>
-
-#include <sys/types.h>
-#include <stdint.h>
-#include <stdbool.h>
-#include <string.h>
-#include <assert.h>
-#include <debug.h>
-#include <errno.h>
-#include <unistd.h>
-
-#include <arch/board/board.h>
-
 #include <drivers/device/spi.h>
 #include <drivers/drv_mag.h>
 #include <drivers/drv_device.h>
 
 #include "hmc5883.h"
-#include <board_config.h>
 
 #ifdef PX4_SPIDEV_HMC
